muokattu delete + delete dokumentaatio

+++ TicketGuru/TicketGuru-Backend/src/main/java/com/example/TicketGuru/web/RestEventController.java
@@ -6,6 +6,7 @@ import java.time.LocalDate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -65,8 +66,12 @@ public class RestEventController {
 	
 	//DELETE
 	@DeleteMapping("/events/{id}")
-	void deleteEvent(@PathVariable Long id) {
+	ResponseEntity<String> deleteEvent(@PathVariable Long id) {
 		log.info("Delete event by id");
+		if (eventRepo.findById(id).isEmpty()) {
+			return ResponseEntity.notFound().build();
+		}
 		eventRepo.deleteById(id);
+		return ResponseEntity.ok("Poistettu");
 	}
 }
